backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.27
psycopg[binary]==3.2.4
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2 
python-jose[cryptography]==3.3.0